+++ tests/integration/nautobot-populate.py
@@ -187,6 +187,7 @@ vlans = [
         "vlan_group": test_vlan_group.id,
         "status": {"name": "Active"},
     },
+    {"name": "Test VLAN 600", "vid": 600, "status": {"name": "Active"}},
 ]
 created_vlans = make_nautobot_calls(nb.ipam.vlans, vlans)
 
@@ -595,6 +596,18 @@ custom_fields = [
     },
 ]
 created_custom_fields = make_nautobot_calls(nb.extras.custom_fields, custom_fields)
+
+###############
+# v2.2+ items #
+###############
+if nautobot_version > version.parse("2.1"):
+    # Create Teams
+    teams = [{"name": "My Test Team"}]
+    created_teams = make_nautobot_calls(nb.extras.teams, teams)
+
+    # Create Contacts
+    contacts = [{"name": "My Contact"}, {"name": "My Contact 2"}]
+    created_contacts = make_nautobot_calls(nb.extras.contacts, contacts)
 
 if ERRORS:
     sys.exit("Errors have occurred when creating objects, and should have been printed out. Check previous output.")